Move Functions to Funcs.py
Clean up functions mat_update and getsets to new file
