+++ CProps.py
@@ -1,16 +1,6 @@
 import bpy
-#Functions for the allmaterials EnumProperty
-def mat_update(self, context):
-    pass
-def getsets(self, context):
-    sets = []
-    mats = bpy.data.materials
+from .Funcs import *
 
-    for mat in mats:
-        sets.append((mat.name, mat.name, mat.name))
-    if not sets:
-        sets = [('None', 'None', 'None')]
-    return sets
 
 bpy.types.Scene.saved_location = bpy.props.FloatVectorProperty(
     name="Pizza_Tool_Saved_Location",
